wall slide & jump feature
The new WallJumpState requires knowledge about its surroundings,
that's why it requires the host to have a top and bottom 'sensor'.
These 'sensors' allow the WallJumpState to correctly predict when it is
right to cling to walls.

Once inside the state, the character will slowly slide down the wall,
which is realized by using a custom friction value while falling.
Clinging to a wall will reset dash and double jump cooldowns.
Dashing while clinging to a wall is no problem and jumping off is
possible as well, with the small restriction of a short input delay
after jumping off, in order to avoid wall jump 'spam'.

The placeholder animation for wall clinging requires a small offset,
different to the one used in other animations. This is why I added a
new animation called 'reset_pose', which is called before an animation
is played. This set a refactor in motion which moved most of the
AnimationState code into the Character.play_animation method.
(pending AnimationState rework)